+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support for all kinds of keys types that `json.dumps` supports, for `Dict`
+  fields.
 
 ## [0.3.2] - 2021-12-07
 ### Fixed